requirements.txt
numpy
pandas
seaborn
scikit-learn
scipy
statsmodels
pyyaml
setuptools
# twine
# wheels
dvc
dvc-gdrive
